+++ lib/R5RS.scm
@@ -433,7 +433,7 @@
 
 ;; -----------------------------------------------------------------------------
 ;; generate Math functions with documentation
-(define _maths (list "exp" "log" "sin" "cos" "tan" "asin" "acos" "atan" "atan"))
+(define _maths (list "log" "sin" "cos" "tan" "asin" "acos" "atan" "atan"))
 
 ;; -----------------------------------------------------------------------------
 (define _this_env (current-environment))
@@ -450,6 +450,17 @@
                                   " math operation (it call JavaScript Math)." name
                                   " function."))
         (iter (cdr fns)))))
+
+;; -----------------------------------------------------------------------------
+(define (exp x)
+  (typecheck "exp" x "number")
+  (if (string=? x.__type__ "complex")
+      (let* ((re (real-part x))
+             (im (imag-part x))
+             (factor (Math.exp re)))
+         (make-rectangular (* factor (cos im))
+                           (* factor (sin im))))
+       (Math.exp x)))
 
 ;; -----------------------------------------------------------------------------
 (define (modulo a b)